Fix SFT training: use bf16 instead of fp16 for Qwen2.5 bfloat16 model
Qwen2.5-7B-Instruct loads in bfloat16 by default; fp16 AMP scaler
fails with '_amp_foreach_non_finite_check_and_unscale_cuda not implemented for BFloat16'.

+++ train/sft_tooltune.py
@@ -327,7 +327,7 @@ def finetune(base_model: str, task_files: list[str], output_dir: str) -> None:
         quantization_config=BitsAndBytesConfig(
             load_in_4bit=True,
             bnb_4bit_quant_type="nf4",
-            bnb_4bit_compute_dtype=torch.float16,
+            bnb_4bit_compute_dtype=torch.bfloat16,
             bnb_4bit_use_double_quant=True,
         ),
         device_map="auto",
@@ -360,7 +360,7 @@ def finetune(base_model: str, task_files: list[str], output_dir: str) -> None:
             max_length=2048,
             report_to="none",
             dataset_text_field="text",
-            fp16=True,
+            bf16=True,
         ),
     )
     trainer.train()
